src/main/webapp/app/entities/concierge/delete/concierge-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ConciergeService } from '../service/concierge.service';

import { ConciergeDeleteDialogComponent } from './concierge-delete-dialog.component';

describe('Concierge Management Delete Component', () => {
    let comp: ConciergeDeleteDialogComponent;
    let fixture: ComponentFixture<ConciergeDeleteDialogComponent>;
    let service: ConciergeService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
        TestBed.configureTestingModule({
            imports: [HttpClientTestingModule],
            declarations: [ConciergeDeleteDialogComponent],
            providers: [NgbActiveModal],
        })
            .overrideTemplate(ConciergeDeleteDialogComponent, '')
            .compileComponents();
        fixture = TestBed.createComponent(ConciergeDeleteDialogComponent);
        comp = fixture.componentInstance;
        service = TestBed.inject(ConciergeService);
        mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
        it('Should call delete service on confirmDelete', inject(
            [],
            fakeAsync(() => {
                // GIVEN
                jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

                // WHEN
                comp.confirmDelete(123);
                tick();

                // THEN
                expect(service.delete).toHaveBeenCalledWith(123);
                expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
            })
        ));

        it('Should not call delete service on clear', () => {
            // GIVEN
            jest.spyOn(service, 'delete');

            // WHEN
            comp.cancel();

            // THEN
            expect(service.delete).not.toHaveBeenCalled();
            expect(mockActiveModal.close).not.toHaveBeenCalled();
            expect(mockActiveModal.dismiss).toHaveBeenCalled();
        });
    });
});
